Added missing error messages in command interpreter. Rewrote register_filesystem to be more efficient. Fixed error in register_filesystem where list was not added to properly. Fixed error in init where initrd was overwritten by memory map. Fixed error in initrd where wrong signature was used.

+++ header/defs.h
@@ -2,7 +2,7 @@
 #include "errors.h"
 
 #define CCPVER 0x0100
-#define PROCESS_BASE_ADDRESS 1 << (sizeof(unsigned int)*8)-1
+#define PROCESS_BASE_ADDRESS 1 << (sizeof(size_t)*8)-1
 
 void null_function(void);
