+++ hwdep/i386/hwdefs.h
@@ -2,5 +2,5 @@
 #define TRUE 1
 #define NULL	0
 
-#define KERNEL_HIGH (1 << (sizeof(unsigned int)*8)-1)
+#define KERNEL_HIGH (1 << (sizeof(size_t)*8)-1)
